# server/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZweckOS API",
    "description": "Internal accounting and operations API. Authenticated routes except `/auth/login`, `/auth/register`, `/auth/forgot-password`, `/auth/reset-password`, `/health`, `/docs`, `/openapi.json`.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/api", "description": "Relative to API origin" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": ["transactions"],
        "summary": "List transactions",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["transactions"],
        "summary": "Create transaction",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "List projects",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Create project",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
...
  }
}